Revert "Разбил на модули"
This reverts commit c242878ae0047f3467c488faec7ac6c047eb1b81.
