--- src/remove-film-cards.js
@@ -1,7 +1,0 @@
-export const removeFilmCards = (filmCards) => {
-  for (let i = 0; i < filmCards.length; i++) {
-    let filmCard = filmCards[i];
-
-    filmCard.unrender();
-  }
-};
